Refactor package structure and improve code modularity
Separate core, providers, and utility functionalities into distinct modules to enhance code organization and maintainability. Updated package names to follow a consistent naming convention (`com.jamhour.core`, `com.jamhour.providers`, `com.jamhour.util`) for better clarity and alignment with industry standards.

- Moved core job-related classes (`Job`, `JobBuilder`, `JobVacancyType`, etc.) to the `core` module.
- Moved provider-specific implementations (`AsalTech`, `Foras`, `Harri`, etc.) to the `providers` module.
- Moved utility classes (`HttpUtils`, `LoggingUtils`, `Serializers`, etc.) to the `util` module.
- Updated imports across all files to reflect the new package structure.
- Added `build.gradle.kts` files for each module to define dependencies and configurations.
- Updated `settings.gradle.kts` to include the new modules (`core`, `providers`, `util`).
