+++ build.gradle.kts
@@ -2,7 +2,6 @@ plugins {
     val kotlinVersion = "2.0.20"
 
     kotlin("jvm") version kotlinVersion
-    kotlin("plugin.serialization") version kotlinVersion
     id("com.github.johnrengelman.shadow") version "8.1.1"
 }
 
@@ -14,17 +13,28 @@ repositories {
 }
 
 dependencies {
-    val coroutinesVersion = "1.9.0-RC.2"
-    val serializationVersion = "1.7.2"
-    val jsoupVersion = "1.18.1"
     val kotlinVersion = "2.0.20"
+    val coroutinesVersion = "1.9.0-RC.2"
 
-    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:$coroutinesVersion")
-    implementation("org.jetbrains.kotlinx:kotlinx-serialization-json:$serializationVersion")
-    implementation("org.jsoup:jsoup:$jsoupVersion") // https://mvnrepository.com/artifact/org.jsoup/jsoup
     implementation("org.jetbrains.kotlin:kotlin-stdlib:$kotlinVersion")
-
+    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:${coroutinesVersion}")
     testImplementation(kotlin("test"))
+}
+
+subprojects {
+    apply(plugin = "org.jetbrains.kotlin.jvm")
+
+    repositories {
+        mavenCentral()
+    }
+
+    dependencies {
+        val coroutinesVersion = "1.9.0-RC.2"
+        val serializationVersion = "1.7.2"
+
+        implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:${coroutinesVersion}")
+        implementation("org.jetbrains.kotlinx:kotlinx-serialization-json:${serializationVersion}")
+    }
 }
 
 tasks.test {
