+++ core/build.gradle.kts
@@ -1,0 +1,22 @@
+plugins {
+    kotlin("jvm")
+}
+
+group = "com.jamhour"
+version = "1.0-SNAPSHOT"
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    implementation(project(":util"))
+    testImplementation(kotlin("test"))
+}
+
+tasks.test {
+    useJUnitPlatform()
+}
+kotlin {
+    jvmToolchain(21)
+}
