+++ core/src/main/kotlin/com/jamhour/core/job/JobVacancyType.kt
