+++ util/build.gradle.kts
@@ -1,0 +1,13 @@
+group = "com.jamhour"
+version = "1.0-SNAPSHOT"
+
+dependencies {
+    testImplementation(kotlin("test"))
+}
+
+tasks.test {
+    useJUnitPlatform()
+}
+kotlin {
+    jvmToolchain(21)
+}
